requirements.txt
<<<<<<< HEAD
discord.py
pandas
pyEX
psycopg2
=======
discord.py
pandas
pyEX
psycopg2
>>>>>>> d51912e (test)
sqlalchemy